requirements.txt
pytest==7.4.4
pyyaml==6.0.1
pytest-playwright==0.5.2
pytest-xdist==3.5.0
pytest-html==4.1.1